+++ src/orquestra/sdk/_base/cli/_dorq/_task/_logs.py
@@ -8,7 +8,7 @@ import typing as t
 from pathlib import Path
 
 from orquestra.sdk.schema.configs import ConfigName
-from orquestra.sdk.schema.workflow_run import WorkflowRunId
+from orquestra.sdk.schema.workflow_run import TaskInvocationId, WorkflowRunId
 
 from .. import _arg_resolvers, _dumpers, _repos
 from .._ui import _presenters
@@ -26,6 +26,7 @@ class Action:
         wf_run_repo=_repos.WorkflowRunRepo(),
         config_resolver: t.Optional[_arg_resolvers.WFConfigResolver] = None,
         wf_run_id_resolver: t.Optional[_arg_resolvers.WFRunIDResolver] = None,
+        task_inv_id_resolver: t.Optional[_arg_resolvers.TaskInvIDResolver] = None,
     ):
         # data sources
         self._wf_run_repo = wf_run_repo
@@ -37,6 +38,9 @@ class Action:
         self._wf_run_id_resolver = wf_run_id_resolver or _arg_resolvers.WFRunIDResolver(
             wf_run_repo=wf_run_repo
         )
+        self._task_inv_id_resolver = (
+            task_inv_id_resolver or _arg_resolvers.TaskInvIDResolver(wf_run_repo)
+        )
 
         # output
         self._presenter = presenter
@@ -45,12 +49,16 @@ class Action:
     def on_cmd_call(
         self,
         wf_run_id: t.Optional[WorkflowRunId],
+        task_invocation_id: t.Optional[TaskInvocationId],
+        fn_name: t.Optional[str],
         config: t.Optional[ConfigName],
         download_dir: t.Optional[Path],
     ):
         try:
             self._on_cmd_call_with_exceptions(
                 wf_run_id=wf_run_id,
+                task_invocation_id=task_invocation_id,
+                fn_name=fn_name,
                 download_dir=download_dir,
                 config=config,
             )
@@ -60,6 +68,8 @@ class Action:
     def _on_cmd_call_with_exceptions(
         self,
         wf_run_id: t.Optional[WorkflowRunId],
+        task_invocation_id: t.Optional[TaskInvocationId],
+        fn_name: t.Optional[str],
         config: t.Optional[ConfigName],
         download_dir: t.Optional[Path],
     ):
@@ -69,8 +79,17 @@ class Action:
         resolved_wf_run_id = self._wf_run_id_resolver.resolve(
             wf_run_id, resolved_config
         )
-        logs = self._wf_run_repo.get_wf_logs(
-            wf_run_id=resolved_wf_run_id, config_name=resolved_config
+        resolver_task_inv_id = self._task_inv_id_resolver.resolve(
+            task_inv_id=task_invocation_id,
+            fn_name=fn_name,
+            wf_run_id=resolved_wf_run_id,
+            config=resolved_config,
+        )
+
+        logs = self._wf_run_repo.get_task_logs(
+            wf_run_id=resolved_wf_run_id,
+            task_inv_id=resolver_task_inv_id,
+            config_name=resolved_config,
         )
 
         if download_dir is not None: