+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 🚨 *Breaking Changes*
-
+* `sdk.WorkflowRun.get_logs()` now accepts TaskInvocationID instead of TaskRunID
 * `sdk.WorkflowRun.get_artifacts()` doesn't accept any arguments anymore. Now, it returns all the artifacts produced by the tasks in the workflow.
+
 
 🔥 *Features*
 
@@ -18,7 +19,7 @@
 * Dorq wf submit now properly prompts users for config selection
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry workflow results`. It shows a preview of the workflow run output artifact and can download the result values to the provided directory.
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry workflow logs`.
-
+* New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry task logs`.
 
 🐛 *Bug Fixes*